+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_sender_async.py
@@ -11,7 +11,11 @@ import uamqp
 from uamqp import SendClientAsync, types
 from azure.core.credentials import AzureSasCredential
 
-from .._common.message import ServiceBusMessage, ServiceBusMessageBatch
+from .._common.message import (
+    ServiceBusMessage,
+    ServiceBusMessageBatch,
+    AMQPAnnotatedMessage,
+)
 from .._servicebus_sender import SenderMixin
 from ._base_handler_async import BaseHandler
 from .._common.constants import (
@@ -36,12 +40,15 @@ if TYPE_CHECKING:
 MessageTypes = Union[
     Mapping[str, Any],
     ServiceBusMessage,
-    List[Union[Mapping[str, Any], ServiceBusMessage]]
+    AMQPAnnotatedMessage,
+    List[Union[Mapping[str, Any], ServiceBusMessage, AMQPAnnotatedMessage]],
 ]
 MessageObjTypes = Union[
     ServiceBusMessage,
     ServiceBusMessageBatch,
-    List[ServiceBusMessage]]
+    AMQPAnnotatedMessage,
+    List[Union[ServiceBusMessage, AMQPAnnotatedMessage]],
+]
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -202,7 +209,8 @@ class ServiceBusSender(BaseHandler, SenderMixin):
         Returns a list of the sequence numbers of the enqueued messages.
 
         :param messages: The message or list of messages to schedule.
-        :type messages: ~azure.servicebus.ServiceBusMessage or list[~azure.servicebus.ServiceBusMessage]
+        :type messages: Union[~azure.servicebus.ServiceBusMessage, ~azure.servicebus.AMQPAnnotatedMessage,
+         List[Union[~azure.servicebus.ServiceBusMessage, ~azure.servicebus.AMQPAnnotatedMessage]]]
         :param schedule_time_utc: The utc date and time to enqueue the messages.
         :type schedule_time_utc: ~datetime.datetime
         :keyword float timeout: The total operation timeout in seconds including all the retries. The value must be
@@ -287,9 +295,7 @@ class ServiceBusSender(BaseHandler, SenderMixin):
         )
 
     async def send_messages(
-        self,
-        message: Union[MessageTypes, ServiceBusMessageBatch],
-        **kwargs: Any
+        self, message: Union[MessageTypes, ServiceBusMessageBatch], **kwargs: Any
     ) -> None:
         """Sends message and blocks until acknowledgement is received or operation times out.
 
@@ -297,8 +303,9 @@ class ServiceBusSender(BaseHandler, SenderMixin):
         `ValueError` if they cannot fit in a single batch.
 
         :param message: The ServiceBus message to be sent.
-        :type message: Union[~azure.servicebus.ServiceBusMessage,~azure.servicebus.ServiceBusMessageBatch,
-         list[~azure.servicebus.ServiceBusMessage]]
+        :type message: Union[~azure.servicebus.ServiceBusMessage, ~azure.servicebus.ServiceBusMessageBatch,
+         ~azure.servicebus.AMQPAnnotatedMessage, List[Union[~azure.servicebus.ServiceBusMessage,
+         ~azure.servicebus.AMQPAnnotatedMessage]]]
         :keyword Optional[float] timeout: The total operation timeout in seconds including all the retries.
          The value must be greater than 0 if specified. The default value is None, meaning no timeout.
         :rtype: None
@@ -330,7 +337,9 @@ class ServiceBusSender(BaseHandler, SenderMixin):
             if isinstance(message, ServiceBusMessageBatch):
                 obj_message = message  # type: MessageObjTypes
             else:
-                obj_message = create_messages_from_dicts_if_needed(message, ServiceBusMessage)
+                obj_message = create_messages_from_dicts_if_needed(  # type: ignore
+                    message, ServiceBusMessage
+                )
                 obj_message = transform_messages_to_sendable_if_needed(obj_message)
                 try:
                     batch = await self.create_message_batch()
@@ -339,7 +348,8 @@ class ServiceBusSender(BaseHandler, SenderMixin):
                 except TypeError:  # Message was not a list or generator.
                     trace_message(cast(ServiceBusMessage, obj_message), send_span)
             if (
-                isinstance(obj_message, ServiceBusMessageBatch) and len(obj_message) == 0
+                isinstance(obj_message, ServiceBusMessageBatch)
+                and len(obj_message) == 0
             ):  # pylint: disable=len-as-condition
                 return  # Short circuit noop if an empty list or batch is provided.