+++ sdk/servicebus/azure-servicebus/samples/sync_samples/authenticate_using_azure_sas_credential.py
@@ -17,8 +17,8 @@ try:
     from urllib.parse import quote as url_parse_quote
 except ImportError:
     from urllib import pathname2url as url_parse_quote
-from azure.core.credentials import AzureSasCredential
 
+from azure.core.credentials import AzureSasCredential
 from azure.servicebus import ServiceBusClient, ServiceBusMessage